home script

+++ database/migrations/2020_11_14_062153_malaysia_import_export.php
@@ -22,8 +22,8 @@ class MalaysiaImportExport extends Migration
             $table->string('sitc4');
             $table->string('sitc5');
             $table->string('country');
-            $table->integer('import');
-            $table->integer('export');
+            $table->bigInteger('import');
+            $table->bigInteger('export');
             $table->index(['country']);
         });
     }
